Bluetooth: Host: Move BR headers to subfolder classic
Move BR/EDR header files from "include/zephyr/
bluetooth" to subfolder "include/zephyr/bluetooth/
classic".
